pyproject.toml
[project]
name = "bird-feeder-ai"
version = "0.1.0"
description = "24/7 bird species detection and tracking using Raspberry Pi 5 + Hailo AI HAT+ 2"
requires-python = ">=3.10"

dependencies = [
    # ML / Training
    "torch>=2.0",
    "torchvision>=0.15",
    "pillow>=10.0",
    "scikit-learn>=1.3",
    "matplotlib>=3.7",
    "seaborn>=0.12",
    "tqdm>=4.65",
    "onnx>=1.14",

    # Backend
    "fastapi>=0.110",
    "uvicorn>=0.30",
    "sqlalchemy>=2.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",

    # Camera
    "opencv-python-headless>=4.8",
    "numpy>=1.24",

    # Weather
    "httpx>=0.27",

    # Audio (optional, uncomment to install)
    # "birdnet>=0.2",

    # Hailo (install on Raspberry Pi only via: sudo apt install hailo-all)
    # "hailo-platform",
]

[project.optional-dependencies]
dev = [
    "jupyter>=1.0",
    "ipykernel>=6.0",
    "pytest>=8.0",
    "ruff>=0.4",
]

[tool.ruff]
line-length = 100
target-version = "py310"
